#13 Add CMake function to add subdirectory
This function conditionally adds a subdirectory predicated on whether
the contained target is already defined in the build tree. This is a
necessary addition because some CARMA Library libraries can be built
standalone or depended on by another library. If a depended-on library
is already added to the build tree, the dependent library should not try
to include it.