Add SAE J2735 Angle data element support (#16)
* #15 Add license and initial README

* #15 Add SAE J2735 dependencies and options

This library depends on nholthaus/units.

* #15 Add SAE J2735 library CMakeLists

It follows the standard layout for CARMA Library libraries.

* #15 Add top-level SAE J2735 configuration

Add the option to build the SAE J2735 library in the top-level CARMA
Library CMake configuration.

* #15 Add angle data element support

This commit adds the Angle data element description and the data element
specialization. It also adds a rudimentary unit test that should be
expanded in the future.