Replace add_subdirectory with FetchContent
Adding CARMA Library libraries via add_subdirectory is causing issues because
libraries can be built standalone but they can also be dependencies for other
libraries. When they are both, their target names collide. FetchContent is
able to handle this, so this spike is trying this out for downstream
consuming projects.