Change `libcarma_add_to_build` to use `FetchContent` (#22)
* Replace add_subdirectory with FetchContent

Adding CARMA Library libraries via add_subdirectory is causing issues because
libraries can be built standalone but they can also be dependencies for other
libraries. When they are both, their target names collide. FetchContent is
able to handle this, so this spike is trying this out for downstream
consuming projects.

* #21 Update libcarma_add_to_build function

The function now uses FetchContent to add the specific CARMA Library to
the build tree. The calls to add_subdirectory have been replaced with
calls to libcarma_add_to_build.